requirements.txt
# Web Framework
flask>=3.0.0
flask-restful>=0.3.10
flask-cors>=4.0.0
gunicorn>=21.0.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0
alembic>=1.13.0

# Trading
alpaca-py>=0.21.0

# Market Data
yfinance>=0.2.40
polygon-api-client>=1.14.0

# Sentiment Analysis
praw>=7.7.0
finnhub-python>=2.4.0
nltk>=3.8.0
transformers>=4.38.0
torch>=2.2.0

# Technical Analysis
ta>=0.11.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
requests-cache>=1.2.0
requests-ratelimiter>=0.6.0
redis>=5.0.0

# Notifications
discord-webhook>=1.3.0
firebase-admin>=6.4.0

# Scheduling
apscheduler>=3.10.0
pandas_market_calendars>=4.4.0
pytz>=2024.1

# Security
flask-jwt-extended>=4.6.0
flask-limiter>=3.5.0
bcrypt>=4.1.0

# Monitoring
prometheus-flask-exporter>=0.23.0
prometheus-client>=0.20.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
